pyproject.toml
[tool.poetry]
name = "aiven_take_home_project"
version = "0.1.0"
description = ""
authors = ["Rohan Khanna <rohan.kh@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
kafka-python = "^2.0.2"
psycopg2 = "^2.8.6"
unittest2 = "^1.1.0"
requests = "^2.24.0"
black = "^20.8b1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
